# Faster RNN layers