Use transpose matrices to get faster inference with RNNs (same as with linear layers).

Compiled | Pre | Post
---- | ---- | ----
Yes | 15.1 ms |  10.2 ms
No | 21.3 ms | 17.3 ms


Benchmark:

```python
import time
import mlx.nn as nn
import mlx.core as mx

lstm = nn.LSTM(512, 512)
mx.eval(lstm)

x = mx.random.uniform(shape=(1, 300, 512))

@mx.compile
def fun(x):
    return lstm(x)

for _ in range(5):
    mx.eval(fun(x))

tic = time.time()
for _ in range(100):
    mx.eval(fun(x))
toc = time.time()
print(10 * (toc - tic))
```

Note this is a breaking change.. but I don't think our RNN layers are used much if at all. Probably better to change it now..