@awni 2024-09-16
Using admm in the inner loop gets another millisecond 🚀