@barronalex 2024-09-16
This is really nice!

Just for fun, a custom kernel to get rid of the `mx.split` gets another ~15% faster on my M1 Max:

```
def lstm_custom(ifgo, cell):
    kernel = mx.fast.metal_kernel(
        name="lstm",
        input_names=["ifgo", "cell", "hidden_size"],
        output_names=["hidden_state", "cell_state"],
        header="""
        template <typename T>
        T sigmoid(T x) {
            auto y = 1 / (1 + metal::exp(-metal::abs(x)));
            return (x < 0) ? 1 - y : y;
        }
        """,
        source="""
            uint elem = thread_position_in_grid.x;
            int index = elem / hidden_size * hidden_size * 4 + elem % hidden_size;

            T i = sigmoid(ifgo[index]);
            T f = sigmoid(ifgo[index + hidden_size]);
            T g = metal::precise::tanh(ifgo[index + hidden_size * 2]);
            T o = sigmoid(ifgo[index + hidden_size * 3]);

            cell_state[index] = f * cell[index] + i * g;
            hidden_state[index] = o * metal::precise::tanh(cell_state[index]);
        """,
    )
    out_shape = list(ifgo.shape)
    out_shape[-1] //= 4
    return kernel(
        inputs=[ifgo, cell, out_shape[-1]],
        output_shapes=[out_shape, out_shape],
        output_dtypes=[ifgo.dtype, ifgo.dtype],
        grid=(ifgo.size // 4, 1, 1),
        threadgroup=(256, 1, 1),
        template=[("T", ifgo.dtype)]
    )
```