@barronalex 2024-09-16
I also toyed with parallelizing across layers as this [post](https://developer.nvidia.com/blog/optimizing-recurrent-neural-networks-cudnn-5/) suggests but for just `num_layers=2` it wasn't really any faster.