@awni 2024-09-17
Very cool! That gets it down to 7.67ms in my benchmark!